Refactor TableRebalanceIntegrationTest / HybridClusterIntegrationTest to avoid duplication in test runs
