+++ pinot-integration-tests/src/test/java/org/apache/pinot/integration/tests/TableRebalanceIntegrationTest.java
@@ -64,7 +64,7 @@ import static org.testng.Assert.assertNull;
 import static org.testng.Assert.assertTrue;
 
 
-public class TableRebalanceIntegrationTest extends HybridClusterIntegrationTest {
+public class TableRebalanceIntegrationTest extends BaseHybridClusterIntegrationTest {
   private static String getQueryString(RebalanceConfig rebalanceConfig) {
     return "dryRun=" + rebalanceConfig.isDryRun() + "&preChecks=" + rebalanceConfig.isPreChecks()
         + "&reassignInstances=" + rebalanceConfig.isReassignInstances()
